:memo: Added: Code from three years ago; unsure of its purpose.
Can't remember what this commit actually contains. Probably this is an incomplete implementation of the Plyr HTML5 player, which was supposed to be pushed two years ago.

Please note that this implementation has not undergone recent testing. It is advisable to refrain from using it until further updates are made in the future.